move hook to separate package
