+++ packages/react-aria-components/package.json
@@ -42,6 +42,7 @@
     "@react-aria/color": "3.0.0-beta.33",
     "@react-aria/focus": "^3.17.1",
     "@react-aria/interactions": "^3.21.3",
+    "@react-aria/loading": "3.0.0-alpha.1",
     "@react-aria/menu": "^3.14.1",
     "@react-aria/toolbar": "3.0.0-beta.5",
     "@react-aria/tree": "3.0.0-alpha.1",
